+++ event-store-metric-meters/src/test/java/uk/gov/justice/services/eventstore/metrics/meters/gauges/UnblockedEventStreamsGaugeMeterTest.java
@@ -4,24 +4,16 @@ import static java.util.Optional.empty;
 import static java.util.Optional.of;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static uk.gov.justice.services.core.annotation.Component.EVENT_LISTENER;
 import static uk.gov.justice.services.metrics.micrometer.meters.MetricsMeterNames.UNBLOCKED_EVENT_STREAMS_GAUGE_NAME;
 
 import uk.gov.justice.services.event.buffer.core.repository.metrics.StreamMetrics;
-import uk.gov.justice.services.eventstore.metrics.meters.gauges.StreamMetricsProvider;
-import uk.gov.justice.services.eventstore.metrics.meters.gauges.UnblockedEventStreamsGaugeMeter;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.slf4j.Logger;
 
 @ExtendWith(MockitoExtension.class)
 public class UnblockedEventStreamsGaugeMeterTest {
@@ -29,62 +21,69 @@ public class UnblockedEventStreamsGaugeMeterTest {
     @Mock
     private StreamMetricsProvider streamMetricsProvider;
 
-    @Mock
-    private Logger logger;
-
-    @InjectMocks
-    private UnblockedEventStreamsGaugeMeter unblockedEventStreamsGaugeMeter;
-
     @Test
     public void shouldGetTheCountOfTheTheNumberOfUnblockedEventStreams() throws Exception {
 
         final Integer numberOfUnblockedStreams = 266;
+        final String source = "some-source";
+        final String component = "some-component";
+
+        final UnblockedEventStreamsGaugeMeter unblockedEventStreamsGaugeMeter = new UnblockedEventStreamsGaugeMeter(
+                source,
+                component,
+                streamMetricsProvider
+        );
 
         final StreamMetrics streamMetrics = mock(StreamMetrics.class);
 
-        when(streamMetricsProvider.getMetrics(EVENT_LISTENER)).thenReturn(of(streamMetrics));
-        when(logger.isDebugEnabled()).thenReturn(false);
+        when(streamMetricsProvider.getMetrics(source, component)).thenReturn(of(streamMetrics));
         when(streamMetrics.unblockedStreamCount()).thenReturn(numberOfUnblockedStreams);
 
         assertThat(unblockedEventStreamsGaugeMeter.measure(), is(numberOfUnblockedStreams));
-
-        verify(logger, never()).debug(anyString());
     }
 
     @Test
     public void shouldReturnZeroIfNoMetricsFound() throws Exception {
 
-        when(streamMetricsProvider.getMetrics(EVENT_LISTENER)).thenReturn(empty());
-        when(logger.isDebugEnabled()).thenReturn(false);
+        final String source = "some-source";
+        final String component = "some-component";
+
+        final UnblockedEventStreamsGaugeMeter unblockedEventStreamsGaugeMeter = new UnblockedEventStreamsGaugeMeter(
+                source,
+                component,
+                streamMetricsProvider
+        );
+
+        when(streamMetricsProvider.getMetrics(source, component)).thenReturn(empty());
 
         assertThat(unblockedEventStreamsGaugeMeter.measure(), is(0));
-
-        verify(logger, never()).debug(anyString());
     }
-
-    @Test
-    public void shouldLogNumberOfUnblockedStreamsIfDebugIsEnabled() throws Exception {
-
-        final Integer numberOfUnblockedStreams = 266;
-
-        final StreamMetrics streamMetrics = mock(StreamMetrics.class);
-
-        when(streamMetricsProvider.getMetrics(EVENT_LISTENER)).thenReturn(of(streamMetrics));
-        when(logger.isDebugEnabled()).thenReturn(true);
-        when(streamMetrics.unblockedStreamCount()).thenReturn(numberOfUnblockedStreams);
-
-        assertThat(unblockedEventStreamsGaugeMeter.measure(), is(numberOfUnblockedStreams));
-        verify(logger).debug("Micrometer counting number of unblocked EVENT_LISTENER event streams.");
-        verify(logger).debug("Number of unblocked EVENT_LISTENER event streams: 266");
-    }
-
+    
     @Test
     public void shouldGetTheCorrectMeterName() throws Exception {
+        final String source = "some-source";
+        final String component = "some-component";
+
+        final UnblockedEventStreamsGaugeMeter unblockedEventStreamsGaugeMeter = new UnblockedEventStreamsGaugeMeter(
+                source,
+                component,
+                streamMetricsProvider
+        );
+
         assertThat(unblockedEventStreamsGaugeMeter.metricName(), is(UNBLOCKED_EVENT_STREAMS_GAUGE_NAME));
     }
 
     @Test
     public void shouldGetTheCorrectMeterDescription() throws Exception {
+        final String source = "some-source";
+        final String component = "some-component";
+
+        final UnblockedEventStreamsGaugeMeter unblockedEventStreamsGaugeMeter = new UnblockedEventStreamsGaugeMeter(
+                source,
+                component,
+                streamMetricsProvider
+        );
+
         assertThat(unblockedEventStreamsGaugeMeter.metricDescription(), is("The current number of streams not in error"));
     }
 }